item service is ready

+++ src/main/java/repositories/ItemRepository.java
@@ -5,10 +5,9 @@ import entities.Items;
 import java.util.List;
 
 public interface ItemRepository {
-
   List<Items> get();
 
   void insert(String name, long categoryId, int price, int available);
 
-  boolean delete(Integer id);
+  boolean delete(Long id);
 }
